[D38132.116208] [Support] mapped_file_region: store size as size_t
Differential Revision: https://reviews.llvm.org/D38132?id=116208

+++ llvm/include/llvm/Support/FileSystem.h
@@ -745,7 +745,7 @@ public:
 
 private:
   /// Platform-specific mapping state.
-  uint64_t Size;
+  size_t Size;
   void *Mapping;
 
   std::error_code init(int FD, uint64_t Offset, mapmode Mode);
@@ -763,7 +763,7 @@ public:
 
   ~mapped_file_region();
 
-  uint64_t size() const;
+  size_t size() const;
   char *data() const;
 
   /// Get a const view of the data. Modifying this memory has undefined
